Task B/requirements (1).txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
Pillow>=9.4.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
